requirements.txt
openai
groq 
requests
google-genai
cohere 
python-dotenv